% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{fetch_data}
\alias{fetch_data}
\title{query MySQL database for a specific transaction (deal)
such as "'LBMLT 2006-1'" (NB: internal quotes) and return a tibble with all
fields (by default) or passing var_names = ('fico, dti')}
\usage{
fetch_data(deal, var_names = "*", driver = "MySQL", username = "root",
  password = "", dbname = "dlf", host = "localhost", table = "loans")
}
\description{
query MySQL database for a specific transaction (deal)
such as "'LBMLT 2006-1'" (NB: internal quotes) and return a tibble with all
fields (by default) or passing var_names = ('fico, dti')
}
\examples{
case1 <- fetch_data("'LBMLT 2006-1'")
var_names = ('fico, dti')
case1 <- fetch_data('LBMLT 2006-1', var_names)
deal = "'LBMLT 2006-1'"
case1 <- fetch_data(deal = deal, var_names = ('fico,dti'))
}
